refactor(transformer-engine): Modularize pipeline step operations
Split monolithic transformer engine into focused modules:
- Extract types to `transformerEngineTypes.ts`
- Extract case conversion utilities to `transformerEngineCaseUtils.ts`
- Extract line operations to `transformerEngineLineOperations.ts`
- Extract text operations to `transformerEngineTextOperations.ts`
- Extract data operations to `transformerEngineDataOperations.ts`
- Extract step application logic to `transformerEngineApplyStep.ts`
